Refactor code.
Add markup comments, fix force unwrap, add navigation folders.
