+++ PasswordPicker.xcodeproj/xcuserdata/olegkirsanov.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "5AB36697-DEDE-49C4-9694-A1AD74844403"
+   type = "1"
+   version = "2.0">
+</Bucket>
