+++ PasswordPicker/App/SceneDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  SceneDelegate.swift
+//  PasswordPicker
+//
+//  Created by Oleg Kirsanov on 06.01.2022.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+    
+}
+
