+++ PasswordPicker/Resources/Assets.xcassets/Contents.json
