+++ PasswordPicker/Resources/Info.plist
